pyproject.toml
[project]
name = "mcp-crash-course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=26.1.0",
    "isort>=8.0.0",
    "langchain-mcp-adapters>=0.2.1",
    "langchain[openai]>=1.2.10",
    "langgraph>=1.0.9",
    "python-dotenv>=1.2.1",
]
